+++ contracts/RariFundManager.sol
@@ -446,7 +446,7 @@ contract RariFundManager is Initializable, Ownable {
      * @dev Ideally, we can add the `view` modifier, but Compound's `getUnderlyingBalance` function (called by `RariFundController.getPoolBalance`) potentially modifies the state.
      * @param currencyCode The currency code of the balance to be calculated.
      */
-    function getRawFundBalance(string memory currencyCode) public returns (uint256) {
+    function getRawFundBalance(string memory currencyCode) public fundEnabled returns (uint256) {
         address erc20Contract = _erc20Contracts[currencyCode];
         require(erc20Contract != address(0), "Invalid currency code.");
 
@@ -774,9 +774,9 @@ contract RariFundManager is Initializable, Ownable {
     function withdrawFrom(address from, string[] calldata currencyCodes, uint256[] calldata amounts) external onlyProxy cachePoolBalances returns (uint256[] memory) {
         // Input validation
         require(currencyCodes.length > 0 && currencyCodes.length == amounts.length, "Lengths of currency code and amount arrays must be greater than 0 and equal.");
-        uint256[] memory pricesInUsd = rariFundPriceConsumer.getCurrencyPricesInUsd();
 
         // Manually cache raw fund balance (no need to check if set previously because the function is external)
+        uint256[] memory pricesInUsd = rariFundPriceConsumer.getCurrencyPricesInUsd();
         _rawFundBalanceCache = toInt256(getRawFundBalance(pricesInUsd));
 
         // Make withdrawals