+++ test/4_fund_rebalancer.js
@@ -19,6 +19,9 @@ const RariFundManager = artifacts.require("RariFundManager");
 const RariFundToken = artifacts.require("RariFundToken");
 const RariFundPriceConsumer = artifacts.require("RariFundPriceConsumer");
 
+const IBoostedSavingsVault = artifacts.require("IBoostedSavingsVault");
+const IERC20 = artifacts.require("IERC20");
+
 // These tests expect the owner and the fund rebalancer of RariFundController and RariFundManager to be set to process.env.DEVELOPMENT_ADDRESS
 contract("RariFundManager", accounts => {
   it("should set accepted currencies", async () => {
@@ -74,7 +77,7 @@ contract("RariFundManager", accounts => {
 
 contract("RariFundController, RariFundManager", accounts => {
   it("should deposit to the fund, approve deposits to pools via RariFundController.approveToPool, and deposit to pools via RariFundController.depositToPool", async () => {
-    let fundControllerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundController.at(process.env.UPGRADE_FUND_CONTROLLER_ADDRESS) : RariFundController.deployed());
+    let fundControllerInstance = await RariFundController.deployed();
     let fundManagerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundManager.at(process.env.UPGRADE_FUND_MANAGER_ADDRESS) : RariFundManager.deployed());
     if (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0) RariFundManager.address = process.env.UPGRADE_FUND_MANAGER_ADDRESS;
 
@@ -91,7 +94,7 @@ contract("RariFundController, RariFundManager", accounts => {
 
       // Approve and deposit to pool
       // TODO: Ideally, we add actually call rari-fund-rebalancer
-      await fundControllerInstance.approveToPool(["dYdX", "Compound", "Aave", "mStable"].indexOf(poolName), currencyCode, amountBN, { from: process.env.DEVELOPMENT_ADDRESS });
+      await fundControllerInstance.approveToPool(["dYdX", "Compound", "Aave", "mStable"].indexOf(poolName), currencyCode, poolName === "mStable" ? web3.utils.toBN(2).pow(web3.utils.toBN(256)).subn(1) : amountBN, { from: process.env.DEVELOPMENT_ADDRESS });
       await fundControllerInstance.depositToPool(["dYdX", "Compound", "Aave", "mStable"].indexOf(poolName), currencyCode, amountBN, { from: process.env.DEVELOPMENT_ADDRESS });
 
       // Check new pool balance
@@ -102,7 +105,7 @@ contract("RariFundController, RariFundManager", accounts => {
   });
 
   it("should withdraw half from all pools via RariFundController.withdrawFromPool", async () => {
-    let fundControllerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundController.at(process.env.UPGRADE_FUND_CONTROLLER_ADDRESS) : RariFundController.deployed());
+    let fundControllerInstance = await RariFundController.deployed();
 
     // For each currency of each pool:
     for (const poolName of Object.keys(pools)) for (const currencyCode of Object.keys(pools[poolName].currencies)) {
@@ -123,7 +126,7 @@ contract("RariFundController, RariFundManager", accounts => {
   });
 
   it("should withdraw everything from all pools via RariFundController.withdrawAllFromPool", async () => {
-    let fundControllerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundController.at(process.env.UPGRADE_FUND_CONTROLLER_ADDRESS) : RariFundController.deployed());
+    let fundControllerInstance = await RariFundController.deployed();
     
     // For each currency of each pool:
     for (const poolName of Object.keys(pools)) for (const currencyCode of Object.keys(pools[poolName].currencies)) {
@@ -136,11 +139,29 @@ contract("RariFundController, RariFundManager", accounts => {
       assert(newBalanceOfUnderlying.isZero());
     }
   });
+
+  it("should claim mStable MTA rewards", async () => {
+    let fundControllerInstance = await RariFundController.deployed();
+
+    // Get unclaimed MTA and indexes necessary to claim MTA and check unclaimed MTA > 0
+    let savingsVaultInstance = await IBoostedSavingsVault.at(pools["mStable"].currencies["mUSD"].savingsVaultContract);
+    var data = await savingsVaultInstance.unclaimedRewards.call(fundControllerInstance.address);
+    assert(web3.utils.toBN(data["0"]).gt(web3.utils.toBN(0)));
+
+    // Claim MTA rewards
+    let mtaTokenInstance = await IERC20.at(pools["mStable"].mtaTokenAddress);
+    var mtaBalanceBeforeClaim = await mtaTokenInstance.balanceOf.call(fundControllerInstance.address);
+    await fundControllerInstance.claimMStableRewards(true, data["1"], data["2"]);
+
+    // Check claimed MTA > 0
+    var mtaBalanceAfterClaim = await mtaTokenInstance.balanceOf.call(fundControllerInstance.address);
+    assert(mtaBalanceAfterClaim.gt(mtaBalanceBeforeClaim));
+  });
 });
 
 contract("RariFundController, RariFundManager", accounts => {
   it("should exchange tokens to and from mStable mUSD via RariFundController.mintMUsd and redeemMUsd", async () => {
-    let fundControllerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundController.at(process.env.UPGRADE_FUND_CONTROLLER_ADDRESS) : RariFundController.deployed());
+    let fundControllerInstance = await RariFundController.deployed();
     if (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0) RariFundController.address = process.env.UPGRADE_FUND_CONTROLLER_ADDRESS;
     let fundManagerInstance = await (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0 ? RariFundManager.at(process.env.UPGRADE_FUND_MANAGER_ADDRESS) : RariFundManager.deployed());
     if (parseInt(process.env.UPGRADE_FROM_LAST_VERSION) > 0) RariFundManager.address = process.env.UPGRADE_FUND_MANAGER_ADDRESS;